requirements.txt
Django
psycopg2
Celery
redis
celery[redis]
geopy
djangorestframework
social-auth-app-django